Update .htaccess for reverse proxy configuration, remove SSH public key, and enhance Nuxt configuration for admin route
This commit adds a note in the .htaccess file regarding the automatic management of the reverse proxy for the /admin route by CloudLinux Node.js Selector. It also removes the id_rsa.pub file for security reasons. Additionally, the Nuxt configuration is updated to exclude the /admin route from prerendering and to set up the admin route for Nuxt Studio, ensuring a more streamlined development process.